界面更新
完成定位界面和职位类型界面的更改.

+++ IamKing/IKHomePage/IKJobTypeView.h
@@ -25,6 +25,6 @@
 @property (nonatomic, strong)NSArray *titleArray;
 @property (nonatomic, assign)CGSize buttonSize;
 
-- (void)adjustBottomLine:(NSInteger )index;
+- (void)adjustBottomLine:(CGFloat )index;
 
 @end
